feat: migrate from Mdviewer to Company with multi-org support
