--- backend/src/data-contracts/simulatorserver/swagger.json
@@ -1,1 +1,0 @@
-{"code":"404","type":"Status report","message":"Runtime Error","description":"No matching resource found for given API Request"}
